counter-app/src/components/AdvanceCounter.tsx
// Counter.tsx
import React, { useState } from "react";

interface CounterProps {
  initialCount?: number;
}

const AdvanceCounter: React.FC<CounterProps> = ({ initialCount = 0 }) => {
  const [count, setCount] = useState(initialCount);

  const decrement = () => {
    if (count > 0) {
      setCount(count - 1);
    }
  };

  return (
    <div>
      <button onClick={decrement}>-</button>
      <span>{count}</span>
      <button onClick={() => setCount(count + 1)}>+</button>
    </div>
  );
};

export default AdvanceCounter;
